Reduce # retries to fix ring diverge issue
What is the ring diverge issue?
Ring diverge will happen after a MapUpdate node had a 'hard' failure
(e.g. power outage), the remaining alive node will have different
pictures of the hash ring.

How ring diverge is triggered?
After a node is reported as 'failure', the Message Server (MS) will
simultaneously notify other alive nodes to get prepared and acked for
the ring change. The totol time for the prep and ack phase is 90
seconds. (90s could be a empirical number). If a node can not ack within
90s, it will be marked as the "not ack" node and kicked out from the
alive nodes.

The above logic sounds reasonable, but there is a scenario that a alive
node will be wrongly marked as "not ack" node, and the ring diverge is
triggered.

How to resolve ring diverge issue?
If a remote node had a 'hard' failure, some on-the-fly jobs on alive nodes
(submitted before the failure was noticed by Mupd8) still need to ping/send
events to the failed node with certain number of retries (6 retries). The time
cost of retrying and then finding out the remote node is unreachable is very big,
especially for the 'hard' failure case where detecting a remote host is
unreachable via network needs much more turnaround time than a 'soft' failure (e.g. app
exits). So it is very easy a live node to run out of 90s, just for pinging the
failed node, without doing any actual prep and ack work. The result is
that the alive node is marked as "not ack" node and removed from the
ring by mistake.

One way to make this above scenario looks better is to reduce the number
of 'unnecessary' retries. The 'unnecessary' retry is defined as: after
Mupd8 is sure that a node has been failed and will be removed from ring,
pinging/sending events to that failed node is 'unnecessary'.

A proposal in this change is: during the phase of ring change, before an event
is sent over network, check if its destination is the failed node nor not. If yes,
do not send this event and even without any retry.

In addition, this change also want to reduce the halt time between two
retries.